+++ alan_project/alan/tests.py
@@ -13,14 +13,13 @@ class ScannerMethodTests(TestCase):
 		lex_code = scanner.scanner_analysis(code)
 		self.assertEqual(lex_code, output)
 
-
 	def test_scanner_comment_without_end(self):
 		'''
 			Test code of FUN to make sure scanner can detect lexical error - 
 			comment without end mark '}'.
 		'''
 		scanner = Scanner()
-		output = ['[k, program]', '[i, testcomment]', '[;]', '[chyba, chybi }]']
+		output = ['[chyba, chybí }]']
 		with open('files/test_fun_comment.txt') as f:
 			code = f.read()
 		lex_code = scanner.scanner_analysis(code)
@@ -32,10 +31,7 @@ class ScannerMethodTests(TestCase):
 			Program Double is just a simple code printing double input value.
 		'''
 		scanner = Scanner()
-		output = ['[k, program]', '[i, double]', '[;]', '[k, integer]', 
-		'[i, u]', '[;]', '[k, begin]', '[k, read]', '[(]', '[i, u]', '[)]',
-		'[;]', '[i, u]', '[=]', '[i, u]', '[*]', '[#, 2]', '[;]', '[k, write]',
-		'[(]', '[i, u]', '[)]', '[;]', '[k, end]', '[.]']
+		output = ['[i, variable]', '[*]', '[#, 2]'] 
 		with open('files/fun_double.txt') as f:
 			code = f.read()
 		lex_code = scanner.scanner_analysis(code)
@@ -47,52 +43,8 @@ class ScannerMethodTests(TestCase):
 
 		'''
 		scanner = Scanner()
-		output = ['[k, program]', '[i, doublecz]', '[;]', '[k, integer]', 
-		'[i, u]', '[;]', '[k, begin]', '[k, read]', '[(]', '[i, u]', '[)]',
-		'[;]', '[i, u]', '[=]', '[i, u]', '[*]', '[#, 2]', '[;]', '[k, write]',
-		'[(]', '[i, u]', '[)]', '[;]', '[k, end]', '[.]']
+		output = ['[i, variable]', '[*]', '[#, 2]'] 
 		with open('files/fun_double_cz.txt') as f:
-			code = f.read()
-		lex_code = scanner.scanner_analysis(code)
-		self.assertEqual(lex_code, output)
-
-	def test_scanner_string_program(self):
-		'''
-			Program String which prints 'Hello World!'
-
-		'''
-		scanner = Scanner()
-		output = ['[k, program]', '[i, string]', '[;]', '[k, begin]', 
-		'[k, write]', '[(]', '[t, Hello World!]', '[)]', '[;]', '[k, end]', 
-		'[.]']
-		with open('files/fun_string.txt') as f:
-			code = f.read()
-		lex_code = scanner.scanner_analysis(code)
-		self.assertEqual(lex_code, output)
-
-	def test_scanner_string_program_without_end(self):
-		'''
-			Program String without end mark '.
-
-		'''
-		scanner = Scanner()
-		output = ['[k, program]', '[i, string]', '[;]', '[k, begin]', 
-		'[k, write]', '[(]', "[chyba, chybi ']"]
-		with open('files/test_fun_string_less.txt') as f:
-			code = f.read()
-		lex_code = scanner.scanner_analysis(code)
-		self.assertEqual(lex_code, output)
-
-	def test_scanner_string_program_with_more_ends(self):
-		'''
-			Program String with more end marks.
-
-		'''
-		scanner = Scanner()
-		output = ['[k, program]', '[i, string]', '[;]', '[k, begin]', 
-		'[k, write]', '[(]', '[t, Hello ]', '[i, world]', '[!]',
-		"[chyba, chybi ']"]
-		with open('files/test_fun_string_more.txt') as f:
 			code = f.read()
 		lex_code = scanner.scanner_analysis(code)
 		self.assertEqual(lex_code, output)
@@ -104,18 +56,13 @@ class ScannerMethodTests(TestCase):
 
 		'''
 		scanner = Scanner()
-		output = ['[k, program]', '[i, operators]', '[;]', '[k, integer]', 
-		'[i, u]', '[;]', '[k, begin]', '[k, read]', '[(]', '[i, u]', '[)]',
-		'[;]', '[k, if]', '[i, u]', '[r, ==]', '[#, 2]', '[k, then]', '[i, u]',
-		'[=]', '[i, u]', '[*]', '[#, 2]', '[;]', '[k, if]', '[i, u]', '[r, >]',
-		'[#, 2]', '[k, then]', '[i, u]', '[=]', '[i, u]', '[+]', '[#, 4]',
-		'[;]', '[k, if]', '[i, u]', '[r, !=]', '[#, 42]', '[k, then]', '[i, u]',
-		'[=]', '[i, u]', '[-]', '[#, 0]', '[;]', '[k, if]', '[i, u]', '[r, >=]',
-		'[#, 42]', '[k, then]', '[i, u]', '[=]', '[i, u]', '[/]', '[#, 42]',
-		'[;]', '[k, if]', '[i, u]', '[r, <]', '[#, 42]', '[k, then]', '[i, u]',
-		'[=]', '[i, u]', '[*]', '[#, 5]', '[;]', '[k, if]', '[i, u]', '[r, <=]',
-		'[#, 42]', '[k, then]', '[i, u]', '[=]', '[i, u]', '[*]', '[#, 5]',
-		'[;]', '[k, write]', '[(]', '[i, u]', '[)]', '[;]', '[k, end]', '[.]']
+		output = ['[i, u]', '[r, ==]', '[#, 2]', '[;]', '[i, u]',
+		'[*]', '[#, 2]', '[;]','[i, u]', '[r, >]','[#, 2]', '[;]', '[i, u]',
+		'[+]', '[#, 4]', '[;]', '[i, u]', '[r, !=]', '[#, 42]', '[;]', '[i, u]',
+		'[-]', '[#, 0]', '[;]', '[i, u]', '[r, >=]', '[#, 42]', '[;]', '[i, u]',
+		'[/]', '[#, 42]', '[;]', '[i, u]', '[r, <]', '[#, 42]', '[;]', '[i, u]',
+		'[*]', '[#, 5]', '[;]', '[i, u]', '[r, <=]', '[#, 42]', '[;]', '[i, u]',
+		'[*]', '[#, 5]']
 		with open('files/test_fun_operators.txt') as f:
 			code = f.read()
 		lex_code = scanner.scanner_analysis(code)
@@ -128,10 +75,7 @@ class ScannerMethodTests(TestCase):
 
 		'''
 		scanner = Scanner()
-		output = ['[k, program]', '[i, unknown]', '[;]', '[k, integer]',
-		'[i, u]', '[;]', '[k, begin]', '[k, read]', '[(]', '[i, u]', '[)]',
-		'[;]', '[i, u]', '[=]', '[i, u]', '[chyba, neznami lexem @]', '[#, 2]',
-		'[;]', '[k, write]', '[(]', '[i, u]', '[)]', '[;]', '[k, end]', '[.]']
+		output = ['[i, u]', '[chyba, neznámý lexém @]', '[#, 2]']
 		with open('files/test_fun_unknown.txt') as f:
 			code = f.read()
 		lex_code = scanner.scanner_analysis(code)
@@ -139,17 +83,13 @@ class ScannerMethodTests(TestCase):
 
 	def test_scanner_logical_operators(self):
 		'''
-			Program Unknown to make sure scanner can correctly 
+			Program to make sure scanner can correctly 
 			find logical operators.
 
 		'''
 		scanner = Scanner()
-		output = ['[k, program]', '[i, logic]', '[;]', '[k, integer]', '[i, u]',
-		'[,]', '[i, v]', '[,]', '[i, result]', '[;]', '[k, begin]',
-		'[i, result]', '[=]', '[i, u]', '[&]', '[i, v]', '[;]', '[i, result]',
-		'[=]', '[i, u]', '[|]', '[i, v]', '[;]', '[i, result]', '[=]', '[!]', 
-		'[i, result]', '[;]', '[k, write]', '[(]', '[i, result]', '[)]', '[;]',
-		'[k, end]', '[.]']
+		output = ['[i, u]', '[&]', '[i, v]', '[;]', '[i, u]', '[|]', '[i, v]',
+		'[;]', '[!]', '[i, result]', '[;]', '[(]', '[i, result]', '[)]']
 		with open('files/test_fun_logical_operators.txt') as f:
 			code = f.read()
 		lex_code = scanner.scanner_analysis(code)
@@ -157,54 +97,23 @@ class ScannerMethodTests(TestCase):
 
 	def test_scanner_end_id(self):
 		'''
-			Program Unknown to make sure scanner can find wrong end of code.
+			Program to make sure scanner can correctly stop with the end of code.
 
 		'''
 		scanner = Scanner()
-		output = ['[k, program]', '[i, number]', '[;]', '[k, integer]',
-		'[chyba, za result]']
+		output = ['[i, result]']
 		with open('files/test_fun_end_id.txt') as f:
-			code = f.read()
-		lex_code = scanner.scanner_analysis(code)
-		self.assertEqual(lex_code, output)
-
-	def test_scanner_end_number(self):
-		'''
-			Program Unknown to make sure scanner can find wrong end of code.
-
-		'''
-		scanner = Scanner()
-		output = ['[k, program]', '[i, number]', '[;]', '[k, integer]',
-		'[i, result]', '[;]', '[k, begin]', '[i, result]', '[=]', 
-		'[chyba, za 42]']
-		with open('files/test_fun_end_number.txt') as f:
-			code = f.read()
-		lex_code = scanner.scanner_analysis(code)
-		self.assertEqual(lex_code, output)
-
-	def test_scanner_end_equal(self):
-		'''
-			Program Unknown to make sure scanner can find wrong end of code.
-
-		'''
-		scanner = Scanner()
-		output = ['[k, program]', '[i, number]', '[;]', '[k, integer]',
-		'[i, result]', '[;]', '[k, begin]', '[i, result]',
-		'[chyba, neukonceny program]']
-		with open('files/test_fun_end_equal.txt') as f:
 			code = f.read()
 		lex_code = scanner.scanner_analysis(code)
 		self.assertEqual(lex_code, output)
 
 	def test_scanner_end_not(self):
 		'''
-			Program Unknown to make sure scanner can find wrong end of code.
+			Program to make sure scanner can correctly stop with the end of code.
 
 		'''
 		scanner = Scanner()
-		output = ['[k, program]', '[i, number]', '[;]', '[k, integer]',
-		'[i, result]', '[;]', '[k, begin]', '[i, result]', '[=]',
-		'[chyba, neukonceny program]']
+		output = ['[i, result]', '[r, ==]', '[!]']
 		with open('files/test_fun_end_not.txt') as f:
 			code = f.read()
 		lex_code = scanner.scanner_analysis(code)
@@ -212,13 +121,11 @@ class ScannerMethodTests(TestCase):
 
 	def test_scanner_end_bigger(self):
 		'''
-			Program Unknown to make sure scanner can find wrong end of code.
+			Program to make sure scanner can correctly stop with the end of code.
 
 		'''
 		scanner = Scanner()
-		output = ['[k, program]', '[i, number]', '[;]', '[k, integer]',
-		'[i, result]', '[;]', '[k, begin]', '[i, result]',
-		'[chyba, neukonceny program]']
+		output = ['[i, result]', '[r, >]']
 		with open('files/test_fun_end_bigger.txt') as f:
 			code = f.read()
 		lex_code = scanner.scanner_analysis(code)
@@ -226,14 +133,36 @@ class ScannerMethodTests(TestCase):
 
 	def test_scanner_end_smaller(self):
 		'''
-			Program Unknown to make sure scanner can find wrong end of code.
+			Program to make sure scanner can correctly stop with the end of code.
 
 		'''
 		scanner = Scanner()
-		output = ['[k, program]', '[i, number]', '[;]', '[k, integer]',
-		'[i, result]', '[;]', '[k, begin]', '[i, result]',
-		'[chyba, neukonceny program]']
+		output = ['[i, result]', '[r, <]']
 		with open('files/test_fun_end_smaller.txt') as f:
+			code = f.read()
+		lex_code = scanner.scanner_analysis(code)
+		self.assertEqual(lex_code, output)
+
+	def test_scanner_end_equal(self):
+		'''
+			Program Unknown to make sure scanner can find unknown lexem "=".
+
+		'''
+		scanner = Scanner()
+		output = ['[i, result]', '[chyba, neznámý lexém =]']
+		with open('files/test_fun_end_equal.txt') as f:
+			code = f.read()
+		lex_code = scanner.scanner_analysis(code)
+		self.assertEqual(lex_code, output)
+
+	def test_scanner_not_equal(self):
+		'''
+			Program Unknown to make sure scanner can find unknown lexem "=".
+
+		'''
+		scanner = Scanner()
+		output = ['[i, result]', '[chyba, neznámý lexém =]']
+		with open('files/test_fun_not_equal.txt') as f:
 			code = f.read()
 		lex_code = scanner.scanner_analysis(code)
 		self.assertEqual(lex_code, output)