+++ alan_project/alan/views.py
@@ -45,13 +45,18 @@ def index(request):
 
 def run_scanner(request):
     code = ''
-    lex_code = ''
+    lex_code = []
+    next = True
     if request.method == 'POST':
         if 'fun_code_area' in request.POST and request.POST['fun_code_area']:
             code = request.POST['fun_code_area']
             lex_code = scanner.scanner_analysis(code)
+            if not lex_code:
+                lex_code.append('[chyba, prazdny program]')
+            if any("chyba," in s for s in lex_code):
+                next = False
     return render(request, 'alan/scanner.html', {'code': code,
-                  'lex_code': lex_code})
+                  'lex_code': lex_code, 'next': next})
 
 
 def run_parser(request):